Updated FrameGenerator.generate() to return long instead of int.

+++ jetty-core/jetty-http3/jetty-http3-common/src/main/java/org/eclipse/jetty/http3/MessageFlusher.java
@@ -69,7 +69,7 @@ public class MessageFlusher extends IteratingCallback
 
         Frame frame = entry.frame;
 
-        int generated = generator.generate(accumulator, entry.endPoint.getStream().getId(), frame, this::onGenerateFailure);
+        long generated = generator.generate(accumulator, entry.endPoint.getStream().getId(), frame, this::onGenerateFailure);
         if (generated < 0)
             return Action.SCHEDULED;
